+++ packages/react/src/hooks/useOrderHistoryWebSocket.ts
@@ -12,7 +12,7 @@ import { useWebSocket } from "react-use-websocket/dist/lib/use-websocket.js";
 import { createSignedWebSocketRequest } from "../utils/websocket.js";
 import { useWasmInitialized } from "../wasm.js";
 import { useConfig } from "./useConfig.js";
-import { useStatus } from "./useStatus.js";
+import { useIsIndexed } from "./useIsIndexed.js";
 import { useWalletId } from "./useWalletId.js";
 
 export type UseOrderHistoryWebSocketParameters = {
@@ -24,13 +24,12 @@ export type UseOrderHistoryWebSocketParameters = {
 export function useOrderHistoryWebSocket(parameters: UseOrderHistoryWebSocketParameters = {}) {
     const isWasmInitialized = useWasmInitialized();
     const config = useConfig(parameters);
-    const status = useStatus(parameters);
     const walletId = useWalletId();
-    const { getWebsocketBaseUrl } = config;
+
     const { enabled = true, onUpdate } = parameters;
 
     const { readyState, sendJsonMessage } = useWebSocket(
-        getWebsocketBaseUrl(),
+        config?.getWebsocketBaseUrl() ?? "",
         {
             filter: () => false,
             onMessage(event) {
@@ -53,9 +52,10 @@ export function useOrderHistoryWebSocket(parameters: UseOrderHistoryWebSocketPar
             share: true,
             shouldReconnect: () => true,
         },
-        enabled,
+        enabled && !!config?.getWebsocketBaseUrl(),
     );
 
+    const { data: isIndexed } = useIsIndexed();
     // Subscribe to wallet updates with auth headers
     useEffect(() => {
         // Capture the current (old) wallet id in a local variable
@@ -65,8 +65,10 @@ export function useOrderHistoryWebSocket(parameters: UseOrderHistoryWebSocketPar
             !enabled ||
             !currentWalletId ||
             readyState !== ReadyState.OPEN ||
-            status !== "in relayer" ||
-            !isWasmInitialized
+            !isWasmInitialized ||
+            !config ||
+            !config.state.seed ||
+            !isIndexed
         )
             return;
 
@@ -78,6 +80,7 @@ export function useOrderHistoryWebSocket(parameters: UseOrderHistoryWebSocketPar
 
         const symmetricKey = getSymmetricKey(config);
         const message = createSignedWebSocketRequest(config, symmetricKey, body);
+
         sendJsonMessage(message);
-    }, [enabled, walletId, readyState, status, isWasmInitialized, sendJsonMessage, config]);
+    }, [enabled, walletId, readyState, isWasmInitialized, sendJsonMessage, config, isIndexed]);
 }