+++ packages/react/src/hooks/useWalletWebSocket.ts
@@ -13,7 +13,7 @@ import { useWebSocket } from "react-use-websocket/dist/lib/use-websocket.js";
 import { createSignedWebSocketRequest } from "../utils/websocket.js";
 import { useWasmInitialized } from "../wasm.js";
 import { useConfig } from "./useConfig.js";
-import { useStatus } from "./useStatus.js";
+import { useIsIndexed } from "./useIsIndexed.js";
 import { useWalletId } from "./useWalletId.js";
 
 export type UseWalletParameters = {
@@ -25,13 +25,11 @@ export type UseWalletParameters = {
 export function useWalletWebsocket(parameters: UseWalletParameters = {}) {
     const isWasmInitialized = useWasmInitialized();
     const config = useConfig(parameters);
-    const status = useStatus(parameters);
     const walletId = useWalletId();
-    const { getWebsocketBaseUrl } = config;
     const { enabled = true, onUpdate } = parameters;
 
     const { readyState, sendJsonMessage } = useWebSocket(
-        getWebsocketBaseUrl(),
+        config?.getWebsocketBaseUrl() ?? "",
         {
             filter: () => false,
             onMessage: (event) => {
@@ -49,9 +47,10 @@ export function useWalletWebsocket(parameters: UseWalletParameters = {}) {
             share: true,
             shouldReconnect: () => true,
         },
-        enabled,
+        enabled && !!config?.getWebsocketBaseUrl(),
     );
 
+    const { data: isIndexed } = useIsIndexed();
     // Subscribe to wallet updates with auth headers
     useEffect(() => {
         // Capture the current (old) wallet id in a local variable
@@ -61,8 +60,10 @@ export function useWalletWebsocket(parameters: UseWalletParameters = {}) {
             !enabled ||
             !currentWalletId ||
             readyState !== ReadyState.OPEN ||
-            status !== "in relayer" ||
-            !isWasmInitialized
+            !isWasmInitialized ||
+            !config ||
+            !config.state.seed ||
+            !isIndexed
         )
             return;
 
@@ -74,6 +75,7 @@ export function useWalletWebsocket(parameters: UseWalletParameters = {}) {
 
         const symmetricKey = getSymmetricKey(config);
         const subscriptionMessage = createSignedWebSocketRequest(config, symmetricKey, body);
+
         sendJsonMessage(subscriptionMessage);
-    }, [enabled, walletId, readyState, status, isWasmInitialized, sendJsonMessage, config]);
+    }, [enabled, walletId, readyState, isWasmInitialized, sendJsonMessage, config, isIndexed]);
 }