+++ packages/react/src/hydrate.ts
@@ -3,10 +3,11 @@
 import type { Config } from "@renegade-fi/core";
 import { type State, hydrate } from "@renegade-fi/core";
 import { type ReactElement, createElement, useEffect, useRef, useState } from "react";
+import * as RustUtils from "../renegade-utils/index.js";
 import { WasmContext } from "./wasm.js";
 
 export type HydrateProps = {
-    config: Config;
+    config: Config | undefined;
     initialState?: State | undefined;
     reconnectOnMount?: boolean | undefined;
 };
@@ -21,12 +22,20 @@ export function Hydrate(parameters: React.PropsWithChildren<HydrateProps>) {
     });
 
     // Hydrate for non-SSR
-    if (!config._internal.ssr) onMount();
+    if (config && !config._internal.ssr) onMount();
+
+    useEffect(() => {
+        RustUtils.default().then(() => {
+            setIsInitialized(true);
+            console.log("Backup effect initialized WASM");
+        });
+    }, []);
 
     // Hydrate for SSR
     const active = useRef(true);
     // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
     useEffect(() => {
+        if (!config) return;
         config.utils
             .default()
             .then(() => {