+++ examples/supported-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @renegade-fi/supported-tokens
 
+## 0.0.5
+
+### Patch Changes
+
+- @renegade-fi/node@0.6.9
+
 ## 0.0.4
 
 ### Patch Changes